backend/requirements.txt
fastapi
uvicorn
langchain==0.1.16
langchain-openai==0.0.8
openai==0.28.1
pydantic
python-dotenv
pymongo